Revert "visualize rules as graphs"
