+++ allTests.ts
@@ -13,7 +13,6 @@ import { treeTests } from "./util/treeTest";
 import { actionsTests } from "./uiCommon/ide/actionsTest";
 import { parserlibTests } from "./parserlib/ddTests";
 import { incrTests } from "./core/incremental/ddTests";
-import { uiCommonTests } from "./uiCommon/dl/ddTest";
 
 // TODO: use a real arg parser
 const flags = new Set(process.argv.slice(2));
@@ -34,7 +33,6 @@ const suites: { [name: string]: Suite } = {
   treeTests,
   actionsTests,
   parserlibTests: parserlibTests(writeResults),
-  uiCommonTests: uiCommonTests(writeResults),
 };
 
 try {
