+++ uiCommon/generic/treeFromRecords.ts
